+++ semana19/projeto/labook/src/controller/userController.ts
@@ -45,4 +45,3 @@ export const signup = async (req: Request, res: Response) => {
    }
 }
 
-
